Move meta test to own file
